tools/chinese_grammar/requirements.txt
ebooklib
beautifulsoup4
google-generativeai
streamlit
typing_extensions
